Resolve a number of GCC-generated warnings
The warnings in test47 seem to be a symptom of a larger problem,
i.e., not an issue with the test set code but rather with the GCC
configuration.  Hopefully the switch to LLVM will resolve those.

+++ minix/commands/partition/partition.c
@@ -376,7 +376,7 @@ void distribute(void)
 				if (base > pe->lowsec) {
 					fprintf(stderr,
 	"%s: fixed partition %u is preceded by too big partitions/holes\n",
-						arg0, ((pe - table) - 1) / 2);
+				arg0, ((unsigned int)(pe - table) - 1) / 2);
 					exit(1);
 				}
 				exp= nil;	/* XXX - Extend before? */
